FeedbackUppdatering
Uppdatering av removeItem funktionen

+++ Lab6/Lab6/ShoppingList.h
@@ -42,7 +42,7 @@ void editItem(struct ShoppingList *list);
 void removeItem(struct ShoppingList *list);
 void saveList(struct ShoppingList *list); //implementeras i laboration 7
 void loadList(struct ShoppingList* list); //implementeras i laboration 7
-bool noNegativeNumbersAllowed(int number);
+bool noNegativeNumbersAllowed(float number);
 
 
 #endif
